% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_telegram.R
\name{tidy_telegram}
\alias{tidy_telegram}
\title{Tidy dataframes from telegram bot}
\usage{
tidy_telegram(token)
}
\arguments{
\item{token}{your access token (genereated by botfather)}
}
\description{
This function allows you extract a tidy dataframe from somewhat messy telegram data
}
\examples{
mybot <- tidy_telegram(token = "123456")
}
\keyword{dataframe}
\keyword{tidy}
